--- src/About.jsx
@@ -1,13 +1,0 @@
-
-import React from 'react'
-import Navigation from './Component/Navigation/Navigation'
-function About() {
-   return (
-      <div>
-         hello
-         <Navigation/>
-      </div>
-   )
-}
-
-export default About
