+++ src/Component/Herosection/Welcome/index.js
@@ -1,0 +1,23 @@
+import React from 'react';
+
+
+const Index=() => {
+    return (
+        <div>
+            <h1 className="center">Our Aim</h1>
+            <div className="line-wrapper">
+                <div className="line"></div>
+            </div>
+            <p className="pgraph">Life is tough sometimes but it doesn't remain that way. Things improve! Speaking up
+             can be hard but reaching out is the first step to feeling better. Talking about mental health issues and stresses 
+             may not be easy but it’s a vital part of getting the help you need. Ladies mental support is a 
+            group to support and help ladies going through tough situations and issues. We welcome you as we work 
+            together to overcome mental challenges/illnesses by providing you a path to healing.</p>
+            <div className="btn-wrapper">
+                <a href="/"><div className="btn">Learn more</div></a>
+            </div>
+        </div>
+    )
+}
+
+export default Index;
