+++ src/Component/Home/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Navigation from '../Navigation/index';
+ import Herosection from '../Herosection/index';
+import Note from '../Note/index';
+
+const Index = () => {
+    return (
+        <div>
+            <Navigation /> 
+            <Herosection />
+            <Note/>
+        </div>
+    )
+}
+
+export default Index;
