+++ src/Component/Note/index.css
@@ -1,0 +1,48 @@
+* {
+    box-sizing: border-box;
+}
+
+.note {
+    margin: 0px auto;
+    padding: 20px;
+    padding-top: 200px;
+    height: 100%;
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+    justify-items: center;
+    grid-gap: 20px;
+    column-gap: 50px;
+    max-width: 900px;
+    align-items: center;
+}
+
+.note-section {
+    background-color: #fe5d9f;
+    width: 300px;
+    height: 400px;
+    border: 1px black solid;
+    display: grid;
+    box-shadow: 0px 20px 20px rgba(0, 0, 0, 0.1);
+    padding: 10px;
+    margin: 10px;
+    border-top: 3px solid;
+    text-align: center;
+    transition: 0.2s ease-out;
+}
+
+.note-section:hover {
+    box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.2);
+    transform: scale(0.99) translate(0px, 5px);
+}
+
+.note-img {
+    width: 120px;
+    height: 120px;
+    border-radius: 50%;
+    margin: 0 auto;
+    padding-top: 10px;
+}
+
+.note-paragraph {
+    font-size: 16px;
+}
